+++ src/Epinova.ElasticSearch.Core/Constants.cs
@@ -4,6 +4,7 @@
     {
         public const string EPiServerConnectionStringName = "EPiServerDB";
         public const string TrackingTable = "ElasticTracking";
+        public const string TrackingFieldIndex = "IndexName";
         public const string CommerceProviderName = "CatalogContent";
     }
 }