+++ tests/TestData/ElasticFixtureSettings.cs
@@ -5,10 +5,6 @@ namespace TestData
 {
     public static class ElasticFixtureSettings
     {
-        public const string PluginName = "mapper-attachments";
-        public static string Language;
-        public const string LanguageName = "norwegian";
         public static string IndexName;
-        public static readonly Type IndexType = typeof(IndexItem);
     }
 }